+++ packages/engine/CHANGELOG.md
@@ -1,8 +1,8 @@
-# @red-request/engine
+# @reddb-io/request-engine
 
 ## 0.1.1
 
 ### Patch Changes
 
 - Updated dependencies [61bcc25]
-  - @red-request/core@0.1.1
+  - @reddb-io/request-core@0.1.1